+++ src/components/Bag/classes.js
