+++ src/components/Bag/index.js
